+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 ### 12.24.0
 
+* `[color]` Using structs for color models instead of bare numbers
+* `[color]` Much simpler converting between color models
 * `[color]` Added method `RGB2CMYK` for converting RGB colors to CMYK
 * `[color]` Added method `CMYK2RGB` for converting CMYK colors to RGB
 * `[color]` Added method `RGB2HSL` for converting RGB colors to HSL
 * `[color]` Added method `HSL2RGB` for converting HSL colors to RGB
-* `[color]` Added method `RGBLuminance` for calculating relative luminance for RGB colors
-* `[color]` Added method `HEXLuminance` for calculating relative luminance for HEX colors
+* `[color]` Added method `Luminance` for calculating relative luminance for RGB colors
 * `[color]` Added method `Contrast` for calculating contrast ratio of foreground and background colors
+* `[color]` Method `RGB2HSB` rewritten from scratch and renamed to `RGB2HSV`
+* `[color]` Method `HSB2RGB` rewritten from scratch and renamed to `HSV2RGB`
+* `[color]` Added more usage examples
+
 
 ### 12.23.0